refactor(image): share one Skia ImageUtils across Desktop/iOS/Web
The Desktop (JVM) and iOS (native) ImageUtils were ~470 lines of near-identical
Skia code (org.jetbrains.skia) — only convertHeicToJpeg actually differed
(FFmpeg vs UIImage). Web was an unimplemented TODO stub.

Introduce a `skiaMain` intermediate source set shared by the targets that bundle
skiko — jvm, native (apple), and wasmJs — and move the single Skia ImageUtils
implementation (resize/compress/crop/rotate/getNaturalSize/warpAffine/drawStrokes/
blurBytes/rasterizeSvg + toImageBitmap) there. Android keeps its own
android.graphics actual; only convertHeicToJpeg stays per-platform.

Net effect: ~900 fewer lines of duplication, and the web target gets a real,
synchronous Skia image pipeline for free (skiko is already on the wasm classpath
via Compose; verified it compiles AND runs on wasm). No behavior change on
Desktop/iOS — the shared code is byte-identical to what they already shipped.

Verified: :homebase-api compiles for wasmJs, jvm, and android. iOS is unverified
locally (no Mac) — the moved code is identical to iOS's prior impl, so the risk is
source-set wiring only; confirm in CI/Xcode.
